(v.3.4.3.248)_update_tampilan_profile_pada_instructor_dan_memindahkan_tampilan_profile_dari_Home_Controller_ke_Profile_Controller_serta_memperbaiki_bug_pada_tampilan_dashboard_pada_waktu_record_sesi_menghilang_jika_sudah_mulai_meeting

+++ app/Http/Controllers/StudentController.php
@@ -387,7 +387,7 @@ class StudentController extends Controller
 
         if ($this->is_student()){
 
-            if(Auth::user()->image_profile != 'user.jpg') {
+            if($file && Auth::user()->image_profile != 'user.jpg') {
                 $destinationPath = 'uploads/student/profile/';
                 File::delete($destinationPath . Auth::user()->image_profile);
             }
